메인 컴포넌트 위치변경 및 안쓰는 코드 삭제중
