+++ src/Components/Login/NaverLogin.js
@@ -1,13 +1,11 @@
 import { useEffect } from "react";
 import React, { useRef } from "react";
 import axios from "axios";
-import { getCookie, setCookie } from "../../hooks/CookieHook";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { useState } from "react";
 import Layout from "../../style/Layout";
 import ReactDaumPost from "react-daumpost-hook";
-import useInput from "../../hooks/UseInput";
 import Loading from "../../style/Loading";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
@@ -41,8 +39,10 @@ const NaverLogin = ({ setGetToken, setUserInfo }) => {
       );
       if (data.accessToken) {
         alert("로그인 성공!");
-        setCookie("accessToken", data.accessToken, { path: "/" });
-        setCookie("refreshToken", data.refresh_token, { path: "/" });
+        sessionStorage.setItem("accessToken", data.accessToken, { path: "/" });
+        sessionStorage.setItem("refreshToken", data.refresh_token, {
+          path: "/",
+        });
         navigate("/main");
       }
       setIsLoading(false);
@@ -66,8 +66,10 @@ const NaverLogin = ({ setGetToken, setUserInfo }) => {
       );
       console.log(data.accessToken);
       if (data.accessToken) {
-        setCookie("accessToken", data.accessToken, { path: "/" });
-        setCookie("refreshToken", data.refresh_token, { path: "/" });
+        sessionStorage.setItem("accessToken", data.accessToken, { path: "/" });
+        sessionStorage.setItem("refreshToken", data.refresh_token, {
+          path: "/",
+        });
         alert("가입을 축하드립니다!");
         navigate("/main");
       } else {
