+++ src/style/GlobalStyle.jsx
@@ -7,17 +7,20 @@ const GlobalStyle = createGlobalStyle`
     src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_two@1.0/NanumSquareRound.woff') format('woff');
     font-weight: normal;
     font-style: normal;
+    font-display: swap
 }
 @font-face {
     font-family: 'NanumSquare';
     src: url('https://cdn.rawgit.com/moonspam/NanumSquare/master/nanumsquare.css') format('woff');
     font-weight: normal;
     font-style: normal;
+    font-display: swap
 }
 
 @font-face {
     font-family: 'SpoqaHanSansNeo-Regular';
     src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2108@1.1/SpoqaHanSansNeo-Regular.woff') format('woff');
+    font-display: swap
 }
   *, *::before, *::after {
 font-family: 'NanumSquare';    box-sizing: border-box;
